split up test suite by module
 - not properly converting to fixtures yet,
   shared stuff for now imported from
   commons_for_tests.py
